Fixes and Documentation
Made a single instance of libVLC instead of 2.
Added file detection for the original game content so it shows the user a message instead of just crashing.
Added generic error messaging that autosaves on unknown errors.
Fixed a special case where Gowron quits the holodeck program for you and the game closes.
Scaled the hotspot visualization boxes.
Documentation and code comments.